add edit koi info

+++ src/page/pond/pond-info/index.jsx
@@ -1,9 +1,11 @@
 import React, { useState, useEffect } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import api from "../../../config/axios";
-import { Button, Form, Input, InputNumber } from "antd";
+import { Button, Form, Image, Input, InputNumber, Upload } from "antd";
 import "./index.scss";
 import dayjs from "dayjs";
+import { PlusOutlined } from "@ant-design/icons";
+import uploadFile from "../../../assets/hook/useUpload";
 
 function PondInfo() {
   const { id } = useParams();
@@ -13,7 +15,48 @@ function PondInfo() {
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(null);
   const navigate = useNavigate();
-  const fetchPondAndWaterReport = async () => {
+  const [fileList, setFileList] = useState([]);
+  const [previewImage, setPreviewImage] = useState("");
+  const [previewOpen, setPreviewOpen] = useState(false);
+  const handlePreview = async (file) => {
+    if (!file.url && !file.preview) {
+      file.preview = await getBase64(file.originFileObj);
+    }
+    setPreviewImage(file.url || file.preview);
+    setPreviewOpen(true);
+  };
+
+  const handleChange = ({ fileList: newFileList }) => setFileList(newFileList);
+
+  const uploadButton = (
+    <button
+      style={{
+        border: 0,
+        background: "none",
+        color: "black",
+      }}
+      type="button"
+    >
+      <PlusOutlined />
+      <div
+        style={{
+          marginTop: 8,
+        }}
+      >
+        Upload
+      </div>
+    </button>
+  );
+
+  const getBase64 = (file) =>
+    new Promise((resolve, reject) => {
+      const reader = new FileReader();
+      reader.readAsDataURL(file);
+      reader.onload = () => resolve(reader.result);
+      reader.onerror = (error) => reject(error);
+    });
+
+  const fetchData = async () => {
     setLoading(true);
     try {
       const [pondResponse, waterReportResponse] = await Promise.all([
@@ -30,7 +73,7 @@ function PondInfo() {
   };
 
   useEffect(() => {
-    fetchPondAndWaterReport();
+    fetchData();
   }, []);
 
   const handleDelete = async () => {
@@ -38,6 +81,10 @@ function PondInfo() {
       alert("This pond has fish, you cannot delete it.");
       return;
     }
+    const confirmDelete = window.confirm(
+      "Are you sure you want to delete this pond?\nThis action cannot be undone."
+    );
+    if (!confirmDelete) return;
     try {
       await api.delete(`pond/${pondId}`);
       alert("Pond deleted successfully");
@@ -46,6 +93,7 @@ function PondInfo() {
       console.error("Failed to delete pond:", error);
     }
   };
+
 
   if (loading) return <div>Loading...</div>;
   if (error) return <div>{error}</div>;
@@ -60,6 +108,9 @@ function PondInfo() {
             onFinish={(values) => {
               const updatePond = async () => {
                 try {
+                  const url = await uploadFile(fileList[0].originFileObj);
+                  console.log(url);
+                  values.pondImage = url;
                   await api.put(`pond/${pondId}`, values);
                   alert("Pond updated successfully");
                   navigate("/managerPond");
@@ -73,38 +124,44 @@ function PondInfo() {
             {" "}
             <div className="pond-info-columns">
               <div className="left-column">
+                <h2>Pond Report</h2>
                 <Form.Item label="Name" name="pondName">
                   <Input />
                 </Form.Item>
                 <Form.Item label="Pond Image" name="pondImage">
-                  <Input />
-                  <img
-                    src={pond.pondImage}
-                    alt="pond"
-                    style={{ width: "100%", height: 100 }}
-                  />
+                  <img src={pond.pondImage} alt="pondImage" />
+                  <Upload
+                    // action="https://660d2bd96ddfa2943b33731c.mockapi.io/api/upload"
+                    listType="picture-card"
+                    fileList={fileList}
+                    onPreview={handlePreview}
+                    onChange={handleChange}
+                    className="upload-container"
+                  >
+                    {fileList.length >= 8 ? null : uploadButton}
+                  </Upload>
                 </Form.Item>
-                <Form.Item label="Area" name="area">
+                <Form.Item label="Area: m2" name="area">
                   <InputNumber min={0} />
                 </Form.Item>
-                <Form.Item label="Depth" name="depth">
+                <Form.Item label="Depth: m" name="depth">
                   <InputNumber min={0} />
                 </Form.Item>
               </div>
               <div className="right-column">
-                <Form.Item label="Volume" name="volume">
+                <Form.Item label="Volume: m3" name="volume">
                   <InputNumber min={0} />
                 </Form.Item>
-                <Form.Item label="Drain Count" name="drainCount">
+                <Form.Item label="Drain Count: " name="drainCount">
                   <InputNumber min={0} />
                 </Form.Item>
-                <Form.Item label="Skimmer Count" name="skimmerCount">
+                <Form.Item label="Skimmer Count: " name="skimmerCount">
                   <InputNumber min={0} />
                 </Form.Item>
-                <Form.Item label="Pumping Capacity" name="pumpingCapacity">
+                <Form.Item label="Pumping Capacity: L" name="pumpingCapacity">
                   <InputNumber min={0} />
                 </Form.Item>
-                <Form.Item label="Amount of Fish" name="amountFish">
+                <Form.Item label="Amount of Fish: " name="amountFish">
                   <InputNumber disabled min={0} />
                 </Form.Item>
               </div>
@@ -124,11 +181,28 @@ function PondInfo() {
             </div>
           </Form>
         </div>
+        {previewImage && (
+          <Image
+            className="image-preview"
+            wrapperStyle={{ display: "none" }}
+            preview={{
+              visible: previewOpen,
+              onVisibleChange: (visible) => setPreviewOpen(visible),
+              afterOpenChange: (visible) => !visible && setPreviewImage(""),
+            }}
+            src={previewImage}
+          />
+        )}
 
         <Form
           name="water-report-form"
           className="water-report"
-          initialValues={waterReport}
+          initialValues={{
+            ...waterReport,
+            waterReportUpdatedDate: dayjs(
+              waterReport.waterReportUpdatedDate
+            ).format("YYYY-MM-DD"),
+          }}
           onFinish={async (values) => {
             try {
               const response = await api.put(
@@ -148,47 +222,38 @@ function PondInfo() {
             <Form.Item label="Water Report ID" name="waterReportId">
               <Input disabled />
             </Form.Item>
-            <Form.Item
-              label="Water Report Updated Date"
-              name="waterReportUpdatedDate"
-            >
+            <Form.Item label="Updated Date" name="waterReportUpdatedDate">
               <Input disabled />
             </Form.Item>{" "}
-            <Form.Item
-              label="Water Report Temperature"
-              name="waterReportTemperature"
-            >
+            <Form.Item label="Temperature(°C):" name="waterReportTemperature">
               <InputNumber min={0} />
             </Form.Item>
-            <Form.Item label="Water Report Oxygen" name="waterReportOxygen">
+            <Form.Item label="Oxygen(mg/L):" name="waterReportOxygen">
               <InputNumber min={0} />
             </Form.Item>
-            <Form.Item label="Water Report pH" name="waterReport_pH">
+            <Form.Item label="pH" name="waterReport_pH">
               <InputNumber min={0} max={14} />
             </Form.Item>
-            <Form.Item label="Water Report Hardness" name="waterReportHardness">
+            <Form.Item label="Hardness(ppm):" name="waterReportHardness">
               <InputNumber min={0} />
             </Form.Item>
-            <Form.Item label="Water Report Ammonia" name="waterReportAmmonia">
+            <Form.Item label="Ammonia(ppm):" name="waterReportAmmonia">
               <InputNumber min={0} />
             </Form.Item>
-            <Form.Item label="Water Report Nitrite" name="waterReportNitrite">
+            <Form.Item label="Nitrite(ppm):" name="waterReportNitrite">
               <InputNumber min={0} />
             </Form.Item>
-            <Form.Item label="Water Report Nitrate" name="waterReportNitrate">
+            <Form.Item label="Nitrate(ppm):" name="waterReportNitrate">
               <InputNumber min={0} />
             </Form.Item>
-            <Form.Item
-              label="Water Report Carbonate"
-              name="waterReportCarbonate"
-            >
+            <Form.Item label="Carbonate(ppm):" name="waterReportCarbonate">
               <InputNumber min={0} />
             </Form.Item>
-            <Form.Item label="Water Report Salt" name="waterReportSalt">
+            <Form.Item label="Salt Level(%): " name="waterReportSalt">
               <InputNumber min={0} />
             </Form.Item>
             <Form.Item
-              label="Water Report Carbon Dioxide"
+              label="Carbon Dioxide(ppm):"
               name="waterReportCarbonDioxide"
             >
               <InputNumber min={0} />
